Update to work with GDAL 3, and also fix lots of warnings found with latest C++. The warnings are mostly harmless but we want them cleaned up to find real issues. Actually found a found 'real' problems with the warnings, although these are all pretty minor.

+++ lib/GeoCalCore/Implementation/geotiff_file.cc
@@ -439,7 +439,7 @@ template<> int GeotiffFile::get_tiftag(tiftag_t K) const
 
 template<> blitz::Array<double, 1> GeotiffFile::get_tiftag(tiftag_t K) const
 {
-  uint32_t count;
+  uint16_t count;
   double *data;
   int status = TIFFGetField(tif, K, &count, &data);
   if(status != 1)
